makefile: switched to stow for installation/management of dotfiles
with stow there is only one copy of the files which lives in the repo.
this facilitate edits as there is nothing to "sync" back and forth to
the repo, once I'm satisfied with an edit, I can simply commit and push,
or checkout if I want to go back.